+++ GPTifier/include/serialization/response_to_json.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <json.hpp>
+#include <string>
+
+nlohmann::json response_to_json(const std::string &response);
