+++ GPTifier/src/serialization/parse_response.cpp
@@ -39,22 +39,3 @@ nlohmann::json parse_response(const std::string &response)
 
     return results;
 }
-
-nlohmann::json response_to_json(const std::string &response)
-{
-    nlohmann::json json;
-
-    try {
-        json = nlohmann::json::parse(response);
-    } catch (const nlohmann::json::parse_error &e) {
-        throw std::runtime_error(fmt::format("Failed to parse response: {}", e.what()));
-    }
-
-    if (json.contains("error")) {
-        if (not json["error"].empty()) {
-            throw std::runtime_error(json["error"]["message"]);
-        }
-    }
-
-    return json;
-}
